+++ SharpSource/SharpSource.Test/UnboundedStackallocTests.cs
@@ -1,19 +1,13 @@
 using System.Threading.Tasks;
-using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SharpSource.Diagnostics;
-using SharpSource.Test.Helpers;
+
+using VerifyCS = SharpSource.Test.CSharpCodeFixVerifier<SharpSource.Diagnostics.UnboundedStackallocAnalyzer, SharpSource.Diagnostics.UnboundedStackallocCodeFix>;
 
 namespace SharpSource.Test;
 
 [TestClass]
-public class UnboundedStackallocTests : DiagnosticVerifier
+public class UnboundedStackallocTests
 {
-    protected override DiagnosticAnalyzer DiagnosticAnalyzer => new UnboundedStackallocAnalyzer();
-
-    protected override CodeFixProvider CodeFixProvider => new UnboundedStackallocCodeFix();
-
     [TestMethod]
     public async Task UnboundedStackalloc()
     {
@@ -21,7 +15,7 @@ public class UnboundedStackallocTests : DiagnosticVerifier
 using System;
 
 var len = new Random().Next();
-Span<int> values = stackalloc int[len];";
+Span<int> values = stackalloc int{|#0:[len]|};";
 
         var result = @"
 using System;
@@ -29,8 +23,7 @@ using System;
 var len = new Random().Next();
 Span<int> values = len < 1024 ? stackalloc int[len] : new int[len];";
 
-        await VerifyDiagnostic(original, "An array is stack allocated without checking the length. Explicitly check the length against a constant value");
-        await VerifyFix(original, result);
+        await VerifyCS.VerifyCodeFix(original, VerifyCS.Diagnostic().WithMessage("An array is stack allocated without checking the length. Explicitly check the length against a constant value"), result);
     }
 
     [TestMethod]
@@ -42,7 +35,7 @@ using System;
 var len = new Random().Next();
 Span<int> values = len < 1024 ? stackalloc int[len] : new int[len];";
 
-        await VerifyDiagnostic(original);
+        await VerifyCS.VerifyNoDiagnostic(original);
     }
 
     [TestMethod]
@@ -56,7 +49,7 @@ if (len < 1024) {
     Span<int> values = len < 1024 ? stackalloc int[len] : new int[len];
 }";
 
-        await VerifyDiagnostic(original);
+        await VerifyCS.VerifyNoDiagnostic(original);
     }
 
     [TestMethod]
@@ -72,7 +65,7 @@ if (len < 1024) {
 Span<int> values = len < 1024 ? stackalloc int[len] : new int[len];
 ";
 
-        await VerifyDiagnostic(original);
+        await VerifyCS.VerifyNoDiagnostic(original);
     }
 
     [TestMethod]
@@ -86,7 +79,7 @@ if (1024 > len) {
     Span<int> values = len < 1024 ? stackalloc int[len] : new int[len];
 }";
 
-        await VerifyDiagnostic(original);
+        await VerifyCS.VerifyNoDiagnostic(original);
     }
 
     [TestMethod]
@@ -94,7 +87,7 @@ if (1024 > len) {
     {
         var original = @"System.Span<int> values = stackalloc int[32];";
 
-        await VerifyDiagnostic(original);
+        await VerifyCS.VerifyNoDiagnostic(original);
     }
 
     [TestMethod]
@@ -106,7 +99,7 @@ using System;
 var len = new Random().Next();
 unsafe
 {
-    var v2 = stackalloc int[len];
+    var v2 = stackalloc int{|#0:[len]|};
 }";
 
         var result = @"
@@ -118,8 +111,7 @@ unsafe
     var v2 = len < 1024 ? stackalloc int[len] : new int[len];
 }";
 
-        await VerifyDiagnostic(original, "An array is stack allocated without checking the length. Explicitly check the length against a constant value");
-        await VerifyFix(original, result);
+        await VerifyCS.VerifyCodeFix(original, VerifyCS.Diagnostic().WithMessage("An array is stack allocated without checking the length. Explicitly check the length against a constant value"), result);
     }
 
     [TestMethod]
@@ -129,7 +121,7 @@ unsafe
 const int len = 32;
 System.Span<int> values = stackalloc int[len];";
 
-        await VerifyDiagnostic(original);
+        await VerifyCS.VerifyNoDiagnostic(original);
     }
 
     [TestMethod]
@@ -137,14 +129,13 @@ System.Span<int> values = stackalloc int[len];";
     {
         var original = @"
 var len = 32;
-System.Span<int> values = stackalloc int[len];";
+System.Span<int> values = stackalloc int{|#0:[len]|};";
 
         var result = @"
 var len = 32;
 System.Span<int> values = len < 1024 ? stackalloc int[len] : new int[len];";
 
-        await VerifyDiagnostic(original, "An array is stack allocated without checking the length. Explicitly check the length against a constant value");
-        await VerifyFix(original, result);
+        await VerifyCS.VerifyCodeFix(original, VerifyCS.Diagnostic().WithMessage("An array is stack allocated without checking the length. Explicitly check the length against a constant value"), result);
     }
 
     [TestMethod]
@@ -154,7 +145,7 @@ System.Span<int> values = len < 1024 ? stackalloc int[len] : new int[len];";
 using System;
 
 var len = 32;
-M(stackalloc int[len]);
+M(stackalloc int{|#0:[len]|});
 void M(Span<int> arr) { }";
 
         var result = @"
@@ -164,8 +155,7 @@ var len = 32;
 M(len < 1024 ? stackalloc int[len] : new int[len]);
 void M(Span<int> arr) { }";
 
-        await VerifyDiagnostic(original, "An array is stack allocated without checking the length. Explicitly check the length against a constant value");
-        await VerifyFix(original, result);
+        await VerifyCS.VerifyCodeFix(original, VerifyCS.Diagnostic().WithMessage("An array is stack allocated without checking the length. Explicitly check the length against a constant value"), result);
     }
 
     [TestMethod]
@@ -177,7 +167,7 @@ using System;
 void Outer()
 {
     var len = new Random().Next();
-    void Inner() => (stackalloc int[len]).ToString();
+    void Inner() => (stackalloc int{|#0:[len]|}).ToString();
 }";
 
         var result = @"
@@ -189,8 +179,7 @@ void Outer()
     void Inner() => (len < 1024 ? stackalloc int[len] : new int[len]).ToString();
 }";
 
-        await VerifyDiagnostic(original, "An array is stack allocated without checking the length. Explicitly check the length against a constant value");
-        await VerifyFix(original, result);
+        await VerifyCS.VerifyCodeFix(original, VerifyCS.Diagnostic().WithMessage("An array is stack allocated without checking the length. Explicitly check the length against a constant value"), result);
     }
 
     [TestMethod]
@@ -202,7 +191,7 @@ using System;
 class Test
 {
     private int len = new Random().Next();
-    void Method() => (stackalloc int[len]).ToString();
+    void Method() => (stackalloc int{|#0:[len]|}).ToString();
 }";
 
         var result = @"
@@ -214,8 +203,7 @@ class Test
     void Method() => (len < 1024 ? stackalloc int[len] : new int[len]).ToString();
 }";
 
-        await VerifyDiagnostic(original, "An array is stack allocated without checking the length. Explicitly check the length against a constant value");
-        await VerifyFix(original, result);
+        await VerifyCS.VerifyCodeFix(original, VerifyCS.Diagnostic().WithMessage("An array is stack allocated without checking the length. Explicitly check the length against a constant value"), result);
     }
 
     [TestMethod]
@@ -230,7 +218,7 @@ class Test
     void Method()
     {
         var len = new Random().Next();
-        Span<int> values = stackalloc int[len];
+        Span<int> values = stackalloc int{|#0:[len]|};
     }
 }";
 
@@ -247,7 +235,6 @@ class Test
     }
 }";
 
-        await VerifyDiagnostic(original, "An array is stack allocated without checking the length. Explicitly check the length against a constant value");
-        await VerifyFix(original, result);
+        await VerifyCS.VerifyCodeFix(original, VerifyCS.Diagnostic().WithMessage("An array is stack allocated without checking the length. Explicitly check the length against a constant value"), result);
     }
 }